% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_stars}
\alias{add_stars}
\title{FUNCTION_TITLE}
\usage{
add_stars(p.value, type = "html")
}
\arguments{
\item{p.value}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
