% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{analyses}
\alias{analyses}
\title{Analysis function for EGAP replication.}
\usage{
analyses(DV, treat, covs = NULL, heterogenous = NULL, subset = NULL,
  FE = NULL, cluster = NULL, IPW = NULL, data, model = "lm",
  margin_at = NULL, status = c(TRUE, TRUE, TRUE))
}
\arguments{
\item{DV}{Dependent variable specified as character.}

\item{treat}{Treatment variable(s) specified as character vector.}

\item{covs}{Covariate(s) specified as character vector.}

\item{heterogenous}{Covariate(s) to interact with treatment(s) specified as character vector.}

\item{subset}{character string specifing logical expression for subsetting.}

\item{FE}{Fixed effects covariate(s) specified as character vector.}

\item{cluster}{Covariate for clustered robust standard errors as defined by \code{multiwayvcov::cluster.vcov()} function. Currently only one clustering variable option supported specified as character vector.}

\item{IPW}{Inverse probability weights specified as character vector.}

\item{data}{Data frame which contains all the relevant variables.}

\item{model}{Character string specifying the model to estimate. Currently only "lm" and "logit" is supported. Default is "lm".}

\item{margin_at}{Character string which should be in the format of \code{'var_name = value'}, defaults to NULL (no marginal effects). This calculates the marginal effects of the \code{treat} variable in Logit and Probit models at these particular levels. Takes only binary variables.}

\item{status}{Logical vector of length 3, specifying whether the model was pre-(R)egistered, run in (S)cript and reported in (P)aper respectively.}
}
\value{
List of three objects. \code{estimates} is estimates from the model and corresponding standard errors. \code{stat} is vector of adjusted R squared and number of observations. \code{model_spec} is logical vector of characteristics of the model.
}
\description{
Analysis function for EGAP replication.
}
\examples{
\dontrun{
analyses(DV = "incumbent_share",
         treat = c('civics','infoonly'),
         covs = NULL,
         subset = "goodnews==1 & dosage==1",
         FE = "blid",
         heterogenous = NULL,
         IPW = "ipw",
         cluster = "newid",
         data = benin_admin,
         model = "lm",
         status = c(TRUE,TRUE,TRUE))
         }

}
