% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_table.R
\name{arrange_table}
\alias{arrange_table}
\title{A function for rearranging 2x2 tables}
\usage{
arrange_table(table, swap_rows = FALSE, swap_cols = FALSE)
}
\arguments{
\item{table}{A 2x2 matrix or data frame}

\item{swap_rows}{Swap rows of table?}

\item{swap_cols}{Swap columns of table?}
}
\description{
This function allows for easy row and column swapping of tables.
}
\examples{
arrange_table(a_table,TRUE,FALSE)
}
\keyword{table}
