% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_dfe_report.R
\name{basic_report}
\alias{basic_report}
\title{Basic template for a DfE Research Report}
\usage{
basic_report(number_sections = FALSE, ...)
}
\arguments{
\item{..., keep_tex, latex_engine, citation_package, highlight, fig_caption, md_extensions, template}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\value{
An R Markdown output format.
}
\description{
Basic template for a DfE Research Report
}
