% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdheader.r
\name{birdheader}
\alias{birdheader}
\title{function_name: birdheader}
\usage{
birdheader(birdimg, flag = TRUE)
}
\arguments{
\item{paramname}{character value, optional argument}

\item{paramname}{integer value, optional argument}

\item{docName}{character value, optional argument}
}
\description{
Purpose: add a state bird picture to markdown or latex 
Creation Date: 10-03-2015
Last Modified: Fri Mar 20 00:41:18 2015
}
\examples{
birdheader('katemoss_CK.jpg',flag=TRUE)
}

