% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_operator.R
\name{chatty}
\alias{chatty}
\title{FO: chatty}
\usage{
chatty(f)
}
\description{
This FO can print the first parameter when processing.
}
\examples{
f <- function(x) x^2
chatty(f)(5)

s <- c(3,2,1)
vapply(s, chatty(f), numeric(1))

}
