% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterboot_mfx.R
\name{clusterboot_mfx}
\alias{clusterboot_mfx}
\title{Perform iteration of clustered marginal effects}
\usage{
clusterboot_mfx(.model, .cluster, .variable, .category)
}
\arguments{
\item{.model}{DESCRIPTION.}

\item{.cluster}{DESCRIPTION.}

\item{.variable}{DESCRIPTION.}

\item{.category}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
