% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{compact}
\alias{compact}
\title{Functional: compact}
\usage{
compact(l)
}
\arguments{
\item{l}{The object list.}
}
\description{
This functional can reserve elements not null in list.
}
\examples{
x <- list(a = 1:10,b=NULL,c=letters[1:10])
compact(x)

}
