% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_mplush0.R
\name{correct_mplush0}
\alias{correct_mplush0}
\title{A fixed scaling correction factor HO for lavaan model to mimic Mplus}
\usage{
correct_mplush0(lavaanmodel)
}
\arguments{
\item{lavaanmodel}{a lavaan object}
}
\description{
This function allows you to get fixed scaling correction factor HO for lavaan model to mimic Mplus
}
\examples{
\dontrun{
correct_mplush0(fit1)
}


}
\keyword{lavaan}
\keyword{scale_factor}
