% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_scf.R
\name{correct_scf}
\alias{correct_scf}
\title{A fixed scaling correction factor HO for lavaan model to mimic Mplus}
\usage{
correct_scf(lavaanmodel)
}
\arguments{
\item{lavaanmodel}{a lavaan object}
}
\description{
This function allows you to get fixed scaling correction factor HO for lavaan model to mimic Mplus
}
\examples{
correct.scf(fit1)

}
