% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\docType{data}
\name{dfe_blues}
\alias{dfe_blues}
\title{DfE's blue colour palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
dfe_blues
}
\description{
DfE's blue colour palette
}
\examples{
\dontrun{
pie(rep(1, length(dfe_blues)), col = dfe_blues)
}

}
\keyword{datasets}
