% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{dfe_colours}
\alias{dfe_colours}
\title{Use DfE's colour palette}
\usage{
dfe_colours(colour = NULL, intensity = NULL)
}
\arguments{
\item{colour}{Can select specific colours from c("blue", "red", "orange", "yellow", "green", "purple")}

\item{intensity}{Can select intensities from c(20,40,60,80,100)}
}
\description{
Provides easy to use DfE colour palettes
}
\examples{
\dontrun{
pie(rep(1, length(dfe_colours)), col = dfe_colours)
}

}
