% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_table.R
\name{dfe_tbl}
\alias{dfe_tbl}
\title{DfE table styling}
\usage{
dfe_tbl(df)
}
\arguments{
\item{df}{data.frame object to make into a LaTeX table}
}
\value{
A LaTeX table object.
}
\description{
DfE table styling
}
\examples{
dfe_tbl(head(mtcars)[1:4])

}
