% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weights.R
\name{dist_weights}
\alias{dist_weights}
\title{Create distributional weights}
\usage{
dist_weights(earns, rho = 1.3, summary = TRUE, standardise = FALSE)
}
\arguments{
\item{earns}{Vector of earnings}

\item{rho}{An estimate of the elasticity of the marginal utility of income - default 1.3}

\item{summary}{Return summary}

\item{standardise}{Standardise the weights}
}
\value{
Fill in
}
\description{
Create distributional weights
}
