% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_operator.R
\name{dot_every}
\alias{dot_every}
\title{FO: dot_every}
\usage{
dot_every(n, f)
}
\arguments{
\item{n}{The n times operation.}

\item{f}{The base fucntion.}
}
\description{
This FO can print a dot when every n times operation.
}
\examples{
x <- lapply(1:100, dot_every(10, runif))
}
