% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ESAWC21_raster_vignette.R
\name{download_ESAWC21_raster_vignette}
\alias{download_ESAWC21_raster_vignette}
\title{Download ESA WorldCover 2021 data for package vignette}
\usage{
download_ESAWC21_raster_vignette(
  local_dir = "./data_vignette",
  file_name = "LC_raster_example.tif",
  force = FALSE
)
}
\arguments{
\item{local_dir}{Path to directory where the file should be saved.
If NULL (default), a system cache directory is used.}

\item{file_name}{Desired name for the downloaded file (e.g. "LC_raster.tif").
Must include the ".tif" extension. Defaults to "LC_raster.tif".}

\item{force}{Logical. If TRUE, re-downloads even if the file already exists.}
}
\value{
The full path to the downloaded raster file.
}
\description{
This helper function downloads a land cover raster from ESA WorldCover 2021 to use in the package vignette.
}
\examples{
\dontrun{
download_ESAWC21_raster_vignette(local_dir = "data_vignette/")
}
}
