% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_viirs_raster.R
\name{download_viirs_raster}
\alias{download_viirs_raster}
\title{Download VIIRS Nighttime Lights data
The downloaded file includes data for Europe in 2024.}
\usage{
download_viirs_raster(local_dir = NULL, file_name = "VIIRS.tif", force = FALSE)
}
\arguments{
\item{local_dir}{Path to directory where the file should be saved.
If NULL (default), a system cache directory is used.}

\item{file_name}{Desired name for the downloaded file (e.g. "VIIRS.tif").
Must include the ".tif" extension. Defaults to "VIIRS.tif".}

\item{force}{Logical. If TRUE, re-downloads even if the file already exists.}
}
\value{
The full path to the downloaded raster file.
}
\description{
This helper function downloads VIIRS Nighttime Lights data to use in the calculation of amount of artificial light at night.
It stores it locally, so it needs to be downloaded only once. Data for Europe in 2024 is downloaded and saved.

For more details on the dataset, go to: https://eogdata.mines.edu/products/vnl/
}
\examples{
\dontrun{
TBC
}
}
