% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_output.R
\name{drop1_output}
\alias{drop1_output}
\title{Computes likelihood-ratio test results for a given mdoel.}
\usage{
drop1_output(model_test, ...)
}
\arguments{
\item{model_test}{Name of the model to be tested.}
}
\value{
Tidy table containing likelihood-ratio test results.
}
\description{
This function is a wrapper of stats::drop1() inside gtsummary::tbl_regression.
}
\examples{
# to be included

}
