% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_helper.R
\name{estimate_lm}
\alias{estimate_lm}
\title{OLS Estimator}
\usage{
estimate_lm(frame_df, fit_formula, IPW, robust, col_names, ...)
}
\arguments{
\item{frame_df}{DESCRIPTION.}

\item{fit_formula}{DESCRIPTION.}

\item{IPW}{Inverse probability weights specified as character vector.}

\item{robust}{Logical. Whether to report heteroskedastic robust standard errors. Implemented only for linear models for now.}

\item{col_names}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
OLS Estimator
}
\examples{
# ADD_EXAMPLES_HERE
}
