% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alan.R
\name{extract_alan}
\alias{extract_alan}
\title{This function calculates artificial light at night around a set of coordinates}
\usage{
extract_alan(coords, raster_path, radius_m)
}
\arguments{
\item{coords}{data.frame or matrix with columns lon, lat (WGS84)}

\item{raster_path}{path to tif raster of your choice}

\item{radius_m}{numeric buffer radius in meters}
}
\value{
data.frame with lon, lat, radius_m, alan in natural scale and log scale
}
\description{
Function to extract artificial light at night around a set of coordinates. The units of the output will
be determined by the ALAN units in the input raster.
}
