% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_operator.R
\name{failwith}
\alias{failwith}
\title{FO: failwith}
\usage{
failwith(default = NULL, f, quiet = FALSE)
}
\description{
This FO can return a default value when Error occurs.
}
\examples{
failwith(NA,log)("a")

failwith(NA,log,quiet = T)("a")

}
