% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{fcosttime}
\alias{fcosttime}
\title{FO: fcosttime}
\usage{
fcosttime(f)
}
\description{
This FO can print the operating time of function before print the outcome.
}
\examples{
fcosttime(delay_by(10,sum))(1:100)

}
