% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fix_join}
\alias{fix_join}
\title{Fix names of variables after join operation}
\usage{
fix_join(.data)
}
\arguments{
\item{.data}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
