% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_table.R
\name{frequency_table}
\alias{frequency_table}
\title{Generate Frequency Function}
\usage{
frequency_table(data)
}
\arguments{
\item{data}{a dataframe}

\item{file}{the filename you want to save}
}
\description{
This function allows you to save file compatible for MPlus
}
\examples{
\dontrun{
  frequency_table(mtcars)
}

}
\keyword{frequency}
\keyword{table}
