% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fround.R
\name{fround}
\alias{fround}
\title{Take a numeric and return its rounded value with set number of decimal points.}
\usage{
fround(x, digits)
}
\arguments{
\item{x}{Any numeric value.}

\item{digits}{Integer. Number of decimal points to report.}
}
\value{
Numeric vector.
}
\description{
Take a numeric and return its rounded value with set number of decimal points.
}
\examples{
fround(x = 1/3,
       digits = 3)

}
