% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infl.R
\name{get_infl}
\alias{get_infl}
\title{Fetch Outturn and OBR forecast of inflation measures}
\usage{
get_infl(infl_measure, base_year = 2020, add_yrs = 30, lng_run = NULL)
}
\arguments{
\item{infl_measure}{Inflation measure; "CPI", "RPI" or "Deflator"}

\item{base_year}{Base year for indexing}

\item{add_yrs}{Number of additional years in series}

\item{lng_run}{Long-run rate, unless specified defaults to be 2\% for CPI and GDP Deflator and 3\% for RPI}
}
\value{
A `data.frame` containing an inflation series
}
\description{
Fetch Outturn and OBR forecast of inflation measures
}
\examples{
get_infl("CPI", base_year = 2020, add_yrs = 0, lng_run = 2)

}
