% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_mode}
\alias{get_mode}
\title{# get_mode}
\usage{
get_mode(x)
}
\arguments{
\item{x}{(vector) A vector}
}
\description{
Get the mode of a vector.
}
\examples{
x <- c(1,2,3,4,5,5,5)
get_mode(x)  # returns 5
}
