% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{\%!in\%}
\alias{\%!in\%}
\title{# %!in% (not-in)}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{(string) vector or NULL: the values to be matched.}

\item{table}{(string) ector or NULL: the values to be matched against.}
}
\description{
Opposite of %in%
}
\examples{
librastall_github("georg-un/usefulR")
}
