% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remark_presentation_format.R
\name{knitr_options_html}
\alias{knitr_options_html}
\title{Knitr options for an HTML output format}
\usage{
knitr_options_html(fig_width, fig_height, fig_retina, keep_md, dev = "png")
}
\value{
An list that can be passed as the \code{knitr} argument of the
  \code{\link{output_format}} function.
}
\description{
Define knitr options for an R Markdown output format that creates
HTML output.
}
\seealso{
\link{knitr_options}, \link{output_format}
}

