% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librastall.R
\name{librastall}
\alias{librastall}
\title{librastall (library() + install.packages())}
\usage{
librastall(package)
}
\arguments{
\item{package}{(string) The name of the library to load and install}
}
\description{
Import a library and install it if it is not already installed.
}
\examples{
librastall("dplyr")
}
