% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librastall.R
\name{librastall_github}
\alias{librastall_github}
\title{Load and Install from Github}
\usage{
librastall_github(repo)
}
\arguments{
\item{repo}{The name of the repo to load and install (string)}
}
\description{
\strong{libra}ry() + in\strong{stall}_github()
Import a library and install it, if it is not already installed.
}
\examples{
librastall_github("georg-un/usefulR")
}
