% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{linebreak2}
\alias{linebreak2}
\title{LaTeX Helper to Create a Cell Linebreak by Symbol Length (Excluding Math)}
\usage{
linebreak2(x, align = c("l", "c", "r"), double_escape = F, symbol_width = 11)
}
\arguments{
\item{x}{DESCRIPTION.}

\item{align}{DESCRIPTION.}

\item{double_escape}{DESCRIPTION.}

\item{symbol_width}{DESCRIPTION.}
}
\value{

}
\description{
LaTeX Helper to Create a Cell Linebreak by Symbol Length (Excluding Math)
}
\examples{

}
