% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load my favorite packages}
\usage{
load_packages(data.mining = FALSE, graphics = FALSE)
}
\arguments{
\item{data.mining}{logical value, optional argument}

\item{graphics}{logical value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}
\examples{
load_packages(data.mining = TRUE)
}

