% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_helper}
\alias{loo_helper}
\title{Leave-m-out estimation}
\usage{
loo_helper(
  data,
  var,
  loo_n = 1,
  loo_fun = function(dat) estimatr::tidy(estimatr::lm_robust(data = dat, formula = mpg ~
    1)),
  parallel = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{var}{character vector of variable(s) name(s) to use for LOO}

\item{loo_n}{number of unique values of var to leave out at a time}

\item{loo_fun}{estimation function}

\item{parallel}{whether to use parallel in \code{plyr::alply}}
}
\description{
Leave-m-out estimation
}
