% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lsPackages.r
\name{lsPackages}
\alias{lsPackages}
\title{function_name: lsPackages}
\usage{
lsPackages(ret = FALSE, pr = TRUE, lib = FALSE,
  ignore.packages = "base", kntr = FALSE)
}
\arguments{
\item{ret}{logical value, optional argument}

\item{pr}{logical value, optional argument}

\item{lib}{logical value, optional argument}

\item{kntr}{logical value, optional argument}
}
\description{
Purpose: lists the packages loaded since the start of the R session
Creation Date: 06-03-2015
Last Modified: Fri Mar  6 16:06:36 2015
}
\examples{
lsPackages(FALSE,TRUE,TRUE)
}

