% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_df}
\alias{make_df}
\title{FUNCTION_TITLE}
\usage{
make_df(.matrix, .names = c("estimate", "std.error", "p.value", "term"))
}
\arguments{
\item{.matrix}{DESCRIPTION.}

\item{.names}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
