% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParent_addPackages.r
\name{makeParent_addPackages}
\alias{makeParent_addPackages}
\title{packages for LaTeX}
\usage{
makeParent_addPackages(package.list = NULL, pass.package = NULL,
  use.defaults = TRUE)
}
\arguments{
\item{all_packges}{logical value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}

