% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParent_header.r
\name{makeParent_header}
\alias{makeParent_header}
\title{Keep the parent document simple}
\usage{
makeParent_header(makeParent_filename, creator = NULL)
}
\arguments{
\item{parentDir}{character value, optional argument}

\item{docName}{character value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}
\examples{
makeParent_header(filename,creator)
}

