% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCsummary.R
\name{MCMCsummary}
\alias{MCMCsummary}
\title{The cleanest summary for mcmc.list object}
\usage{
MCMCsummary(mcmcObj)
}
\arguments{
\item{mcmcObj}{mcmc.list object}
}
\description{
A function to report the summary table of mcmc object
}
