% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_custom_losses.R
\name{mean_squared_euclid_dist}
\alias{mean_squared_euclid_dist}
\title{Mean Squared Euclidean Distance}
\usage{
mean_squared_euclid_dist(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (tensor)}

\item{y_pred}{Predictions (tensor of the same shape as y_true)}
}
\description{
A custom keras loss function which calculates the mean suqared euclidean distance between predicted and true values.
Useful for multi-target regression where the target variable is a point in n-dimensional space.
}
\examples{
\dontrun {
# Create custom metric
metric_mean_squared_euclid_dist <- custom_metric("mean_squared_euclid_dist", function(y_true, y_pred) {
    return( mean_squared_euclid_dist(y_true, y_pred) )
})

# Compile model with custom loss & custom metric
model \%>\% compile(
    optimizer = adam(),
    loss = mean_squared_euclid_dist,
    metrics = c('mse', metric_mean_squared_euclid_dist)
)
}
}
