% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multiple pattern replacement.}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{Character vector containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible.}

\item{replacement}{A replacement for matched pattern in sub and gsub. Coerced to character if possible.}

\item{x}{A character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.}

\item{...}{Other parameters passed to \code{gsub()}}
}
\value{
Numeric vector.
}
\description{
Multiple pattern replacement.
}
\examples{
\dontrun{
   mgsub(pattern = c("_", " ["),
         replacement = c("\\\\_", paste("", .latex_spacing, "[")),
         x = x,
         fixed = TRUE)
         }

}
