% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpost.R
\name{newpost}
\alias{newpost}
\title{Create a blogdown post quickly}
\usage{
newpost(lang = "en")
}
\arguments{
\item{lang}{the language of the post}
}
\description{
This will trash any non-master branch installs, and spend time re-installing the packages
}
\keyword{lavaan}
\keyword{scale_factor}
