% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{not-in}
\alias{not-in}
\alias{\%!in\%}
\title{Not-In}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\description{
Opposite of \%in\%
}
\examples{
\dontrun{
x <- 1
x \%!in\% c(2,3,4,5)  # TRUE
}
}
