% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocme_mod.R
\name{ocme_mod}
\alias{ocme_mod}
\title{FUNCTION_TITLE}
\usage{
ocme_mod(w, rev.dum = TRUE, vcov = NULL)
}
\arguments{
\item{w}{DESCRIPTION.}

\item{rev.dum}{DESCRIPTION.}

\item{vcov}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
