% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_relaibility.R
\name{omega_reliability}
\alias{omega_reliability}
\title{Omega Reliability Function}
\usage{
omega_reliability(lavaanobject)
}
\arguments{
\item{lavaanobject}{A Lavaan Object}
}
\description{
This function allows you to Calculate Omega Reliaility from Lavaan
}
\examples{
omega_reliability(model2)
omega_reliability(model1)
omega_reliability(fa_bi)
omega_reliability(fa1)
omega_reliability(fa3)
}
