% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfround.R
\name{pfround}
\alias{pfround}
\title{Take a numeric and return its rounded value with set number of decimal points as a character.}
\usage{
pfround(x, digits)
}
\arguments{
\item{x}{Any numeric value.}

\item{digits}{Integer. Number of decimal points to report.}
}
\value{
Character vector.
}
\description{
Take a numeric and return its rounded value with set number of decimal points as a character.
}
\examples{
pfround(x = 1/3,
        digits = 3)

}
