% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{pm}
\alias{pm}
\title{Plus or minus}
\usage{
pm(a, b)
}
\arguments{
\item{a}{The base, or point estimate value.}

\item{b}{The value to be added and subtracted. The margin of error.}
}
\description{
Defines a plus or minus operator. Useful for interval formation.
}
\examples{
pm(2,4)
}
