% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmc_corMat.R
\name{ppmc_corMat}
\alias{ppmc_corMat}
\title{PPMC graphic matrix for correlation matrix}
\usage{
ppmc_corMat(simData.list, data)
}
\arguments{
\item{simData.list}{a list including all simulated data}

\item{data}{the original data}
}
\description{
This function allows you to get fixed scaling correction factor HO for lavaan model to mimic Mplus
}
\examples{
simData.list <- ppmc_simData(model.Posterior.all, iternum)
ppmc_corMat(simData.list, data)


}
\keyword{lavaan}
\keyword{scale_factor}
