% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmc_simData.R
\name{ppmc_simData}
\alias{ppmc_simData}
\title{Simulate data based on posterior distribution}
\usage{
ppmc_simData(model.Posterior.all, iternum, data)
}
\arguments{
\item{model.Posterior.all}{the posterior distribution generated by R2jags}

\item{iternum}{number of interation for simulating data}
}
\value{
simData.list
}
\description{
This function allows you to get fixed scaling correction factor HO for lavaan model to mimic Mplus
}
\examples{
ppmc_simData(model.Posterior.all, iternum)

}
\keyword{lavaan}
\keyword{scale_factor}
