% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_init.R
\name{proj_init}
\alias{proj_init}
\title{Initialize a new project based on a standardized R framework}
\usage{
proj_init(name, dir, umask = "002", use_renv = TRUE, use_git = TRUE, ...)
}
\arguments{
\item{name}{the name of the project.}

\item{dir}{the directory in which to create the project.}

\item{umask}{sets the umask to the specified value.}

\item{use_renv}{if \code{TRUE} initializes renv for the project.}

\item{use_git}{if \code{TRUE} initializes git for the project.}

\item{...}{additional arguments.}
}
\description{
Initialize a new project based on a standardized R framework.
Creates a project directory with structure;\preformatted{supplied_name  
├── config
├── data
├── logs
├── R
├── src
├── tests
├── run.R
├── README.md
├── supplied_name.Rproj
└── .gitignore
}

By default, Will also initialize the project using git & {renv}, with the
option to disable this functionality.
}
