% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_nanodrop.R
\name{quant_nanodrop}
\alias{quant_nanodrop}
\title{Calculates nanodrop paramters for RNA}
\usage{
quant_nanodrop(data)
}
\arguments{
\item{A}{data frame created with read_nanodrop}
}
\value{
A data frame with nanodrop paramaters calculated
}
\description{
Calculates nanodrop paramters for RNA
}

