% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_it.R
\name{read_it}
\alias{read_it}
\title{Interactively select and read multiple filetypes using default options.}
\usage{
read_it(...)
}
\arguments{
\item{...}{further arguments to be passed.}
}
\value{
a data.frame containing the data which has been read.
}
\description{
Interactively select (utilising {file.choose()}) and read
multiple filetypes using default options. The function reads the file
and prints the function used to do so. The printed function can be
used to replace the call to {read_it()} to be used in reproducible
code.
}
