% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nanodrop.R
\name{read_nanodrop}
\alias{read_nanodrop}
\title{Import raw Nanodrop file}
\usage{
read_nanodrop(file)
}
\arguments{
\item{file}{A .xlsx or .xls file-path (nanodrop export must be converted from .xml)}
}
\value{
A data frame with absorbance and wavelength data
}
\description{
Import raw Nanodrop file
}
