% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remark_template.R
\name{remark_template}
\alias{remark_template}
\title{function_name: remark_template.R}
\usage{
remark_template(toc_depth = 3, template = "default", css = NULL,
  fig_width = 5, fig_height = 5, fig_retina = if (!fig_caption) 2,
  fig_caption = FALSE, dev = "png", smart = TRUE, self_contained = TRUE,
  theme = "default", highlight = "default", mathjax = "default",
  includes = NULL, keep_md = FALSE, lib_dir = NULL,
  md_extensions = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{fig_width}{numeric value, optional argument (default 5)}

\item{fig_height}{numeric value optional argument (default 5)}

\item{dev}{character argument (default pdf)}

\item{highlight}{character value, optional argument (default = "default")}

\item{includes}{= NULL, optional argument}

\item{pandoc_args}{= NULL, optional argument}

\item{keep_tex}{logical value, optional argument}

\item{paramname}{integer value, optional argument}
}
\description{
Purpose: Provides a framework for a basic remark presentation
Creation Date: 20-01-2016
Last Modified: Tue Feb  9 12:14:09 2016
}

