% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_format.R
\name{research_format}
\alias{research_format}
\title{function_name:}
\usage{
research_format(rootDir = getwd(), fig_width = 5, fig_height = 5,
  use_tikz = FALSE, fig_crop = TRUE, dev = "pdf", highlight = "default",
  keep_tex = FALSE, includes = NULL, pandoc_args = NULL,
  knitrSetup = TRUE)
}
\arguments{
\item{fig_width}{numeric value, optional argument (default 5)}

\item{fig_height}{numeric value optional argument (default 5)}

\item{fig_crop}{logical value, optional argument (default TRUE)}

\item{dev}{character argument (default pdf)}

\item{highlight}{character value, optional argument (default = "default")}

\item{keep_tex}{logical value, optional argument}

\item{includes}{= NULL, optional argument}

\item{pandoc_args}{= NULL, optional argument}

\item{paramname}{integer value, optional argument}
}
\description{
Purpose: 
Creation Date: 10-11-2015
Last Modified: Tue Nov 10 11:50:29 2015
}

