% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_dfe_report.R
\name{research_report}
\alias{research_report}
\title{Bookdown style template for a DfE Research Report}
\usage{
research_report(...)
}
\value{
An R Markdown output format with bookdown features e.g. cross-referencing.
}
\description{
Bookdown style template for a DfE Research Report
}
