% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.link.R
\name{rev.link}
\alias{rev.link}
\title{Inverse Link Function for logistic regression}
\usage{
\method{rev}{link}(logofOdds)
}
\arguments{
\item{logofOdds}{log of odds ratio}
}
\description{
This function automate calculating reverse link function value for logistic regression
}
\author{
Jihong Zhang
}
\keyword{Function}
\keyword{Link}
