% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_link.R
\name{rev_link}
\alias{rev_link}
\title{Inverse Link Function for logistic regression}
\usage{
rev_link(logofOdds)
}
\arguments{
\item{logofOdds}{log of odds ratio}
}
\description{
This function automate calculating reverse link function value for logistic regression
}
\author{
Jihong Zhang
}
\keyword{Function}
\keyword{Link}
