% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_code.R
\name{reverse_code}
\alias{reverse_code}
\title{Reverse Coding Function}
\usage{
reverse_code(variable)
}
\arguments{
\item{variable}{A numerical vector}
}
\description{
This function allows you to reverse coding your vector (Max to Min)
}
\examples{
reverse_code()
}
\keyword{cats}
