% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2html.R
\name{rmd2html}
\alias{rmd2html}
\title{function_name: rmd2html.R}
\usage{
rmd2html(filename)
}
\arguments{
\item{filename}{character value, the file to convert from Rmd to HTML}
}
\description{
Purpose: 
Creation Date: 09-09-2015
Last Modified: Wed Sep  9 17:24:15 2015
}
\examples{
rmd2html("lecture6.Rmd")

}

