% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2rscript.R
\name{rmd2rscript}
\alias{rmd2rscript}
\title{## script for converting .Rmd files to .R scripts.
#### Kevin Keenan 2014}
\usage{
rmd2rscript(infile)
}
\arguments{
\item{infile}{The file path needed to be transformed
The function is quite simplisting in that it reads a .Rmd file and adds
comments to non-r code sections, while leaving R code without comments
so that the interpreter can run the commands.}
}
\description{
This function will read a standard R markdown source file and convert it to
an R script to allow the code to be run using the "source" function.
}
