% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsl.R
\name{rsl}
\alias{rsl}
\title{rsl: run, save, load}
\usage{
rsl(func, run = TRUE, obj_name, d)
}
\arguments{
\item{func}{Method to be run}

\item{run}{Boolean value. If TRUE, run the method in func, return the result, and save the result in directory d as obj_name.rda.
If FALSE, load obj_name.rda from directory d and return it.}

\item{obj_name}{Name to assign to the output of func and of the .rda object that will be saved in the directory d.}

\item{d}{directory to where obj_name.rda will be saved and loaded from.}
}
\value{
returns the output of whatever is chosen for func
}
\description{
Wrapper to handle executing a method and saving or loading the output all in one instance.
Gives two options when running a function:
\enumerate{
\item Run the function and save the result as a .rda file in the specified directory
\item Look into the specified directory and load the object instead of running the function
}
}
