% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list_df.R
\name{save_list_df}
\alias{save_list_df}
\title{Write dataframes saved in a list}
\usage{
save_list_df(objectlist, path)
}
\arguments{
\item{objectlist}{The list contains several dataframes as its elements.}

\item{path}{A string of path, ended by "\\\" in windows, like "D:\\\"}
}
\value{
Write csv files in the certain path and return NULL.
}
\description{
\code{save_list_df} writes dataframes saved in a list into certain folder in batch.
}
