% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrap_image_table.r
\name{scrape_image_data}
\alias{scrape_image_data}
\title{function_name: scrape_image_table}
\usage{
scrape_image_data(url, table)
}
\arguments{
\item{url}{character value, optional argument}

\item{table}{integer value, optional argument}
}
\description{
Purpose: get urls for images from table, potentially save images
Creation Date: 10-03-2015
Last Modified: Fri Mar 20 00:42:11 2015
}
\examples{
scrape_image_data('url',2)
}

