% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dir.R
\name{set_dir}
\alias{set_dir}
\title{Attempts to set working directory to the location where the script
containing the function call is saved.}
\usage{
set_dir(quietly = FALSE)
}
\arguments{
\item{quietly}{when \code{FALSE} the outcome and working directory are
printed.}
}
\description{
Attempts to set working directory to the location where
the script containing the function call is saved.
}
