% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rsl_default.R
\name{set_rsl_default}
\alias{set_rsl_default}
\title{Default settings for rsl}
\usage{
set_rsl_default(run_default = TRUE, override_default = FALSE, d_default = NULL)
}
\arguments{
\item{run_default}{boolean value: the default setting to be used for "run" argument in rsl function. see rsl documentation for details.}

\item{override}{hidden argument for rsl function. If true, rsl will use run_default and d_default regardless of input
values for its "run" and "d" arguments.}

\item{d}{the default argument to be used for "d" in runfast}
}
\value{
does not return anything
}
\description{
Sets the default settings for the "run" argument of rsl. You can
also choose if this default argument overides manually imputed arguments for "run".
}
