% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seed.R
\name{set_seed}
\alias{set_seed}
\title{Wrapper for set.seed function in R.}
\usage{
set_seed(.seed = 12345, .parallel = FALSE)
}
\arguments{
\item{.seed}{The number for the random number generation.}

\item{.parallel}{Logical. Whether to set seed for parallel computation.}
}
\value{
Nothing.
}
\description{
Wrapper for set.seed function in R.
}
\examples{
\dontrun{
set_seed(.seed = 12345,
         .parallel = FALSE)
         }

}
