% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimpleEduData}
\alias{SimpleEduData}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{item1}{min1, max6}
  \item{item2}{min1, max6}
  \item{item3}{min1, max6}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
SimpleEduData
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
