% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{srt2subtitle}
\alias{srt2subtitle}
\title{srt2subtitle}
\usage{
srt2subtitle(infilepath, outfilepath)
}
\arguments{
\item{infilepath}{The infilepath like "D://data//BBC.srt".}

\item{outfilepath}{The outfilepath like "D://data//BBC.txt"}
}
\value{
write file into certain path
}
\description{
This function can convert srt to subtitle of txt format
}
