% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_dfe_report.R
\name{start_dfe_report}
\alias{start_dfe_report}
\title{Initiate an rmarkdown DfE templated report}
\usage{
start_dfe_report()
}
\arguments{
\item{title}{}
}
\description{
Initiate an rmarkdown DfE templated report
}
