% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_glm.R
\name{summarise_glm}
\alias{summarise_glm}
\title{Multiple diagnostics for a lm/glm or gam from mgcv package}
\usage{
summarise_glm(glmobject, xcoord = 1, xlab = "x", ylab, main = "")
}
\arguments{
\item{glmobject}{a lm, glm or gam{mgcv} object.}

\item{xcoord}{The independent variable that will be used to plot against the fitted values.
Can be a vector of same length as dependent variable or an integer between 1 and the number of predictors used in the model.
The default value is the first predictor used in the formula of the glm.}

\item{xlab}{Specify the x label for plotting the fitted values. Default is "x" or the column name if xcoord comes from a dataframe.
Note: use data[,x] format to keep column name accessible to this function, not data$name or data[[x]].}

\item{ylab}{Specify ylabel manually. Default is the name response variable in the glm}

\item{main}{Title of fitted values plot as a character or string. Default is no title..}
}
\value{
does not return anything.
}
\description{
Generates multiple diagnostics of a lm/glm/gam object from other base functions.
Prints the summary, residual plots, a QQ plots, the acf of the residuals and a plot the model fit over the raw data.
The dependent variable for the fitted values plot can be specified as one of the predictors used in the glm, or any vector
or data frame column of correct length.
}
