% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_fun.R
\name{table_fun}
\alias{table_fun}
\title{Generate markdown or LaTeX table from list of analyses() function outputs.}
\usage{
table_fun(.table_list, .row_names, .col_names, .title,
  .type = getOption("usefulr.type", c("html", "latex", "markdown")),
  .pad = getOption("usefulr.html_pad", 0),
  .col_print = getOption("usefulr.html_col", 5),
  .latex_colwidth = getOption("usefulr.latex_colwidth",
  ">{\\\\centering\\\\arraybackslash\\\\hsize=.5\\\\hsize}X"),
  .latex_scalebox = getOption("usefulr.latex_scalebox", 1),
  .latex_size = getOption("usefulr.latex_size", "small"),
  .latex_spacing = getOption("usefulr.latex_spacing", "\\\\hspace{0.2in}"),
  .latex_sanitize = getOption("usefulr.latex_sanitize", function(str) {    
  mgsub(pattern = c("_", " ["), replacement = c("\\\\_", paste("",
  .latex_spacing, "[")), x = str, fixed = TRUE) }),
  .latex_floating = list(TRUE, "table"),
  .latex_placement = getOption("usefulr.latex_placement", "H"))
}
\arguments{
\item{.table_list}{List of estimate objects generated by analyses() function.}

\item{.row_names}{Names of rows in output table.}

\item{.col_names}{Names of models in list of analyses() outputs.}

\item{.title}{Character string specifying the name of the table.}

\item{.type}{"html","markdown" or "latex" as specified in knitr::kable function.}

\item{.pad}{Integer number for output table padding.}

\item{.col_print}{Number of columns to split by. Defaults to 5.}

\item{.latex_colwidth}{If .type = "latex", character string specifying column width and alignment. Defaults to ">{\\centering\\arraybackslash\\hsize=.5\\hsize}X".}

\item{.latex_scalebox}{If .type = "latex", numeric value for latex table rescaling. Defaults to 0.9.}

\item{.latex_size}{If .type = "latex", character string specifying font size for latex table. Defaults to "footnotesize".}

\item{.latex_spacing}{If .type = "latex", character string specifying the latex code to be inserted between estimate and standard error in table. Defaults to "\\hspace{0.2in}"}

\item{.latex_sanitize}{If .type = "latex", function specifying the text sanitizing function for table.}

\item{.latex_floating}{If type="latex", first element of list, logical, defines whether the latex table will be floating, while the second element of list defines the floating environment. Defaults to list(TRUE, "table").}

\item{.latex_placement}{If type="latex", character string specifying table floating placement. Defaults to "H".}
}
\value{
Markdown or LaTeX table of estimated models.
}
\description{
Generate markdown or LaTeX table from list of analyses() function outputs.
}
\examples{
\dontrun{
table_fun(.table_list = table3,
          .row_names = c("Constant","Info Only","Civics",
                         "Treatment","Private","Public",
                         "High Dosage x Civics","Observations",
                         "Adj. R-squared"),
          .col_names = c("Low Dose", "High Dose\\\\\\nNo Civics",
                         "High Dose\\\\\\nNo Civics", "All\\\\\\nHigh Dose",
                         "All\\\\\\nCommunes"),
          .type = "html",
          .title = "Replication of Table 3")
          }

}
