% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_results.R
\name{table_results}
\alias{table_results}
\title{Outputs a table with model results}
\usage{
table_results(label_list, model_test, test)
}
\arguments{
\item{label_list}{List of predictor names to appear in the table of results.}

\item{model_test}{Name of the model to produce a result table from.}

\item{test}{either 'F' or 'Chisq'}
}
\value{
A table with model results. It can be easily saved in html to include in your reports / manuscript.
}
\description{
This function takes a model run using \code{lme4} and outputs a table of results, including model coefficients and likelihood-ratio test of each predictor.
}
\examples{
# to be included

}
