% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.mat.r
\name{tex.mat}
\alias{tex.mat}
\title{function_name: tex.mat}
\usage{
tex.mat(X.mat, matrix_name = NULL, round.val = 6, endofline = TRUE,
  color.pos.neg = TRUE)
}
\arguments{
\item{paramname}{character value, optional argument}

\item{paramname}{integer value, optional argument}
}
\description{
Purpose: Write better LaTeX matrices
Creation Date: 06-03-2015
Last Modified: Fri Mar  6 10:37:23 2015
}
\examples{
tex.mat(matrix(1:4,nrow=2),'X')
}

