\name{TeX_hw_title}
\alias{TeX_hw_title}
\title{Set knitr up the way I like it}
\usage{
TeX_hw_title(semester = "", course = "", instructor = "", writers = "",
  contact = "", title = "", due = "", assignment = "", location = "")
}
\arguments{
  \item{course}{character value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}
\examples{
knitrSetup(use.tikz = FALSE)
}

