% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github.R
\name{update_github}
\alias{update_github}
\title{Update all packages from github}
\usage{
update_github()
}
\description{
This will trash any non-master branch installs, and spend time re-installing the packages
}
\keyword{lavaan}
\keyword{scale_factor}
