% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_dt.R
\name{frequency_table}
\alias{frequency_table}
\alias{\%!in\%}
\title{Useful functions to manipulate datasets.}
\usage{
frequency_table(data)

x \%!in\% y
}
\arguments{
\item{data}{a dataframe}

\item{x}{the value to be tested}

\item{y}{the vector to be tested}
}
\description{
Generate Frequency Function
}
\section{Functions}{
\itemize{
\item \code{frequency_table()}: frequency_table is a function to summarize data

\item \code{x \%!in\% y}: %!in% means x not in y. Similar to !%in%

}}
\examples{
\dontrun{
  frequency_table(mtcars)
}

}
\keyword{frequency}
\keyword{table}
