% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulR-package.R
\docType{package}
\name{usefulR-package}
\alias{usefulR}
\alias{usefulR-package}
\title{usefulR: Useful R Packages}
\description{
Collection of useful R packages.
}
\author{
\strong{Maintainer}: Georg Unterholzner

}
\keyword{internal}
