% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovCL.R
\name{vcov_cluster}
\alias{vcov_cluster}
\title{FUNCTION_TITLE}
\usage{
vcov_cluster(.model, .cluster = NULL, .adjust = NULL)
}
\arguments{
\item{.model}{DESCRIPTION.}

\item{.cluster}{DESCRIPTION.}

\item{.adjust}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
