% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_C.r
\name{verbatim.paste}
\alias{verbatim.paste}
\title{Compile C code}
\usage{
verbatim.paste(txt)
}
\arguments{
\item{filename}{character value, optional argument}

\item{flags}{character value, optional argument}
}
\description{
Compiles C code and run it, formatting results from system
}

