% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{where}
\alias{where}
\title{Logical functional: where}
\usage{
where(f, x)
}
\arguments{
\item{f}{A function returns logical value.}

\item{x}{The list.}
}
\description{
Return a logical vector to judge every elements in list.
}
\examples{
df <- data.frame(x = 1:3, y = c("a","b","c"))
where(is.factor,df)

}
