% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mplus_csv.R
\name{write_mplus_csv}
\alias{write_mplus_csv}
\title{Save CSV file compatible for Mplus}
\usage{
write_mplus_csv(data, file)
}
\arguments{
\item{data}{a dataframe}

\item{file}{the filename you want to save}
}
\description{
\code{write_mplus_csv} allows you to save file compatible for MPlus. Specifically,
headers will be removed.
}
\examples{
write_mplus_csv(mtcars, file="~/Desktop/mtcars.csv")

}
\seealso{
This function wraps up \code{\link[utils]{write.table}}.
}
\keyword{csv}
\keyword{mplus}
