% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_mean.R
\name{wtd_mean}
\alias{wtd_mean}
\title{Compute weighted mean.}
\usage{
wtd_mean(x, weights = NULL, normwt = "ignored", na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{A numeric vector of weights.}

\item{normwt}{Specify normwt=TRUE to make weights sum to length(x) after deletion of NAs.}

\item{na.rm}{Set to FALSE to suppress checking for NAs}
}
\value{
Returns scalar vector of length one.
}
\description{
Compute weighted mean.
}
\examples{
\dontrun{set.seed(1)
         x <- runif(500)
         wts <- sample(1:6, 500, TRUE)
         mn <- wtd.mean(x = x, weights = wts)}

}
