% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_mean.R
\name{wtd_sd}
\alias{wtd_sd}
\title{Compute weighted standard deviation.}
\usage{
wtd_sd(x, weights = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{A numeric vector of weights.}

\item{na.rm}{Set to FALSE to suppress checking for NAs}
}
\value{
Returns scalar vector of length one.
}
\description{
Compute weighted standard deviation.
}
\examples{
\dontrun{set.seed(1)
         x <- runif(500)
         wts <- sample(1:6, 500, TRUE)
         mn <- wtd_mean(x = x, weights = wts)
         sd <- wtd_sd(x = x, weights = wts)}

}
