% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{analyses}
\alias{analyses}
\title{Analysis Function for Replications.}
\usage{
analyses(DV, treat, data, model = "lm", covs = NULL,
  heterogenous = NULL, subset = NULL, FE = NULL, cluster = NULL,
  IV_list = NULL, robust = is.null(cluster), ri = NULL, IPW = NULL,
  treat_only = FALSE, margin_at = NULL, status = NULL,
  stars = FALSE, round_digits = 3, return_df = FALSE)
}
\arguments{
\item{DV}{Dependent variable specified as character.}

\item{treat}{Treatment vector of variables specified as character vector.}

\item{data}{Data frame which contains all the relevant variables.}

\item{model}{Character string specifying the model to estimate. Currently only "lm" (OLS), "logit" (Binary Logit), "probit" (Binary Probit), "ologit" (Ordered Logit) and "oprobit" (Ordered Probit) models are supported. Default is "lm".}

\item{covs}{Character vector of covariates specified as character vector.}

\item{heterogenous}{Character vector of covariates to interact with treatments specified as character vector.}

\item{subset}{character string specifing logical expression for subsetting.}

\item{FE}{Character vector of fixed effects covariates specified as character vector.}

\item{cluster}{Covariate for clustered robust standard errors as defined by \code{multiwayvcov::cluster.vcov} function. Currently only one clustering variable option supported specified as character vector.}

\item{IV_list}{Character string. Should be a character string which presents valid IV formula as specified in \code{lfe::felm}.}

\item{robust}{Logical. Whether to report heteroskedastic robust standard errors. Implemented only for linear models for now.}

\item{ri}{Numeric. If not NULL, gives number of iterations to use for calculation of non-parametric Studentized randomization inference p-values from two-sided test. Only implemented for model = "lm" at the moment.}

\item{IPW}{Inverse probability weights specified as character vector.}

\item{treat_only}{Logical vector of length 1, specifying whether only \code{treat} estimates should be reported. Defaults to \code{FALSE}.}

\item{margin_at}{Character string which should be in the format of \code{'var_name = value'}, defaults to NULL (no marginal effects). This calculates the marginal effects of the \code{treat} variable in Logit and Probit models at these particular levels. Takes only binary variables.}

\item{status}{Logical vector of length 3, specifying whether the model was pre-(R)egistered, run in (S)cript and reported in (P)aper respectively.}

\item{stars}{Logical. If \code{FALSE} no stars are passed to printout.}

\item{round_digits}{Integer. How many decimal points to round to in the output.}

\item{return_df}{If \code{TRUE} dataframe used for estimation will be returned.}
}
\value{
List of three objects. \code{estimates} is estimates from the model and corresponding standard errors. \code{stat} is vector of adjusted R squared and number of observations. \code{model_spec} is logical vector of characteristics of the model.
}
\description{
Analysis Function for Replications.
}
\examples{
\dontrun{
usefulr::analyses(DV = "mpg",
                  treat = "vs",
                  model = "lm",
                  covs = c("carb", "disp"),
                  heterogenous = "cyl",
                  IPW = "drat",
                  cluster = "gear",
                  subset = "qsec >= 16",
                  FE = "vs",
                  treat_only = FALSE,
                  status = c(T,T,F),
                  data = mtcars,
                  margin_at = NULL)
         }

}
