% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_table.R
\name{dfe_tbl}
\alias{dfe_tbl}
\title{DfE table styling}
\usage{
dfe_tbl(
  df,
  label = knitr::opts_current$get("label"),
  caption = " ",
  row.names = TRUE,
  ...
)
}
\arguments{
\item{df}{data.frame object to make into a DfE style table}

\item{label}{The table reference label. By default obtained from knitr chunk}

\item{caption}{caption for the table}

\item{row.names}{include row names in table}

\item{...}{Other arguments passed to knitr::kable}
}
\value{
A LaTeX table object in the DfE style. Can also create DfE style-like HTML tables, although support is limited.
}
\description{
Wrapper for knitr::kable to produce tables in the DfE style.
}
\examples{
dfe_tbl(head(mtcars)[1:4])

}
