% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alan.R
\name{extract_alan}
\alias{extract_alan}
\title{This function calculates artificial light at night around a set of coordinates}
\usage{
extract_alan(coords, raster_path, radius_m)
}
\arguments{
\item{coords}{data.frame or matrix with columns lon, lat (WGS84)}

\item{raster_path}{path to terra raster of your choice}

\item{radius_m}{numeric buffer radius in meters}

\item{cat_vals}{integer vector of raster values to consider 'impervious' (only for categorical) (cat_vals = 50 for impervious surface)}

\item{value_scale}{numeric: if fractional values are 0-100 set to 100; default 1 (0-1)}
}
\value{
data.frame with lon, lat, radius_m, percent_impervious (0-100)
}
\description{
Function to extract artificial light at night around a set of coordinates.
}
