% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_land_cover.R
\name{extract_land_cover}
\alias{extract_land_cover}
\title{This function calculate proportion of impervious surface (or any other habitat type) given a set of coordinates}
\usage{
extract_land_cover(coords, radius_m, cat_vals = c(50), value_scale = 100)
}
\arguments{
\item{coords}{data.frame or matrix with columns lon, lat (WGS84)}

\item{radius_m}{numeric buffer radius in meters}

\item{cat_vals}{integer vector of raster values to consider 'impervious' (only for categorical) (cat_vals = 50 for impervious surface)}

\item{value_scale}{numeric: if fractional values are 0-100 set to 100; default 1 (0-1)}

\item{rast}{terra raster}

\item{type}{"categorical"}
}
\value{
data.frame with lon, lat, radius_m, percent_impervious (0-100)
}
\description{
Function to extract the proportion of land cover around a given set of coordinates. 
It uses ESA WorldCover 2021 data. By default, the function loads and uses land cover data from Spain, but the function 
can take data from other geographical areas, passed to the 'rast' argument.

For more information on the land cover data use here, check: https://worldcover2021.esa.int/
}
