\name{knitrSetup}
\alias{knitrSetup}
\title{Set knitr up the way I like it}
\usage{
knitrSetup(rootDir = FALSE, use.tikz = FALSE)
}
\arguments{
  \item{rootDir}{character value, optional argument}

  \item{use.tikz}{logical value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}
\examples{
knitrSetup(use.tikz = FALSE)
}

