% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librastall.R
\name{librastall}
\alias{librastall}
\title{Load and Install}
\usage{
librastall(package)
}
\arguments{
\item{package}{The name of the library to load and install (string)}
}
\description{
\strong{libra}ry() + in\strong{stall}.packages()
Import a library and install it, if it is not already installed.
}
\examples{
librastall("dplyr")
}
