% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librastall.R
\name{librastall_github}
\alias{librastall_github}
\title{librastall_github (library() + install.github())}
\usage{
librastall_github(repo)
}
\arguments{
\item{repo}{(string) The name of the repo to load and install}
}
\description{
Import a library and install it if it is not already installed.
}
\examples{
librastall_github("georg-un/usefulR")
}
