% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/makeParent.r
\name{makeParent}
\alias{makeParent}
\title{Keep the parent document simple}
\usage{
makeParent(parentDir = NULL, docName = "main_document", overwrite = FALSE)
}
\arguments{
\item{parentDir}{character value, optional argument}

\item{docName}{character value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}
\examples{
makeParent(parentDir = '~/courses/cs511/code_bank')
}

