% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeParent_addPackages}
\alias{makeParent_addPackages}
\title{packages for LaTeX}
\usage{
makeParent_addPackages(all_packages = TRUE)
}
\arguments{
\item{all_packges}{logical value, optional argument}
}
\description{
This code sets some of the values for knitr that I like.
}
\examples{
makeParent_addPackages(all_packages = TRUE)
}

