% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lists.r
\name{merge_lists}
\alias{merge_lists}
\title{function_name: merge_lists}
\usage{
merge_lists(lists, append.lists = TRUE, descending.priority = FALSE)
}
\arguments{
\item{lists}{a list of list, optional argument}

\item{append.lists}{logical argument, optional argument}

\item{descending.priority}{logical argument, optional argument}
}
\description{
Purpose: merge two or more lists together
Creation Date: 19-03-2015
Last Modified: Mon May  4 17:48:23 2015
}
\examples{
list1 = list('dogs'=c('chows','beagles'), cows='milk', food = 'hot dogs')
list2 = list('dogs'='doberman', cows=c('milk','beef'), friends = 'nick')
list3 = list('dogs'='puppy', cows=c('barnyard','beef'), girls='peach', friends = 'ian')
usefulR::merge_lists(list(list1,list2,list3))

}

