% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{not-in}
\alias{not-in}
\alias{\%!in\%}
\title{# Not-in}
\arguments{
\item{x}{(string) vector or NULL: the values to be matched.}

\item{table}{(string) ector or NULL: the values to be matched against.}
}
\description{
Opposite of %in%
}
\examples{
\dontrun{
1 \%!in\% c(2,3,4,5)  # TRUE
}
}
