% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_relaibility.R
\name{omega_reliability}
\alias{omega_reliability}
\title{Omega Reliability Function}
\usage{
omega_reliability(lavaanobject)
}
\arguments{
\item{lavaanobject}{a lavaan object using cfa() or sem(). Make sure the scaling method is z-score, marker variable scaling will make this function output different results.}
}
\description{
This function automate omega reliability from lavaan package
}
\examples{
omega_reliability(model2)
omega_reliability(model1)
omega_reliability(fa_bi)
omega_reliability(fa1)
omega_reliability(fa3)
}
\keyword{omega}
\keyword{reliability}
