% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_relaibility.R
\name{omega_reliability}
\alias{omega_reliability}
\title{Omega Reliability Function}
\usage{
omega_reliability(object)
}
\arguments{
\item{object}{The lavaan model object provided after runing the 'cfa', sem, growth, or lavaan functions.}
}
\description{
This function automate calculating omega reliability from lavaan package
}
\examples{
library(dplyr)
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '
fit <- cfa(HS.model, data=HolzingerSwineford1939)
omega_reliability(fit)
}
\keyword{omega}
\keyword{reliability}
