% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_dfe_report.R
\name{research_report}
\alias{research_report}
\title{Test this function}
\usage{
research_report(...)
}
\arguments{
\item{..., keep_tex, latex_engine, citation_package, highlight, fig_caption, md_extensions, template}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\value{
An R Markdown output format.
}
\description{
Test this function
}
