% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_fun.R
\name{table_fun}
\alias{table_fun}
\title{Generate markdown or LaTeX table from list of analyses() function outputs.}
\usage{
table_fun(
  .table_list,
  .row_names,
  .col_names,
  .title,
  .label = NULL,
  .print_status = FALSE,
  .type = getOption("usefulr.type", "markdown"),
  .pad = getOption("usefulr.html_pad", 0),
  .col_print = getOption("usefulr.html_col", 6),
  .group_means = NULL,
  .hypotheses = NULL,
  .add_rows = NULL,
  .add_stars = FALSE,
  .add_modelspec = FALSE,
  .round_digits = 3
)
}
\arguments{
\item{.table_list}{List of estimate objects generated by analyses() function.}

\item{.row_names}{Names of rows in output table.}

\item{.col_names}{Names of models in list of analyses() outputs.}

\item{.title}{Character string specifying the name of the table.}

\item{.label}{Character string specifying the label used to refer to table.}

\item{.print_status}{Logical. Whether to print status of the model.}

\item{.type}{Character string. Either "html","markdown" or "latex" as specified in knitr::kable function.}

\item{.pad}{Integer number for output table padding}

\item{.col_print}{Number of columns to split by. Defaults to 6.}

\item{.group_means}{Matrix of group means with row names. Number of columns must coincide with the main table output}

\item{.hypotheses}{Matrix of directional hypotheses with row names. Number of columns must coincide with the main table output}

\item{.add_rows}{<atrix with other statistics to add to the main output. Number of columns must coincide with the main table output}

\item{.add_stars}{Logical. Whether to add statistical significance stars}

\item{.add_modelspec}{Logical. Whether to add statistical significance stars}

\item{.round_digits}{Numeric. The rounding of report is done up to 10^{-.round_digits}.
param .latex_colwidth If .type = "latex", character string specifying column width and alignment. Defaults to ">{\\centering\\arraybackslash\\hsize=.5\\hsize}X".
param .latex_scalebox If .type = "latex", numeric value for latex table rescaling. Defaults to 0.9.
param .latex_size If .type = "latex", character string specifying font size for latex table. Defaults to "footnotesize".
param .latex_spacing If .type = "latex", character string specifying the latex code to be inserted between estimate and standard error in table. Defaults to "\\hspace{0.2in}"
param .latex_sanitize If .type = "latex", function specifying the text sanitizing function for table.
param .latex_floating If type="latex", first element of list, logical, defines whether the latex table will be floating, while the second element of list defines the floating environment. Defaults to list(TRUE, "table").
param .latex_placement If type="latex", character string specifying table floating placement. Defaults to "H".}
}
\value{
Markdown or LaTeX table of estimated models.
}
\description{
Generate markdown or LaTeX table from list of analyses() function outputs.
}
\examples{
\dontrun{
table_fun(.table_list = table3,
          .row_names = c("Constant","Info Only","Civics",
                         "Treatment","Private","Public",
                         "High Dosage x Civics","Observations",
                         "Adj. R-squared"),
          .col_names = c("Low Dose", "High Dose\\\\\nNo Civics",
                         "High Dose\\\\\nNo Civics", "All\\\\\nHigh Dose",
                         "All\\\\\nCommunes"),
          .type = "html",
          .title = "Replication of Table 3")
          }

}
