% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mplus_csv.R
\name{write_mplus_csv}
\alias{write_mplus_csv}
\title{Write CSV file for Mplus Function}
\usage{
write_mplus_csv(data, file)
}
\arguments{
\item{data}{a dataframe}

\item{file}{the filename you want to save}
}
\description{
This function allows you to save file compatible for MPlus
}
\examples{
write_mplus_csv(mtcars, file="~/Desktop/mtcars.csv")

}
\keyword{csv}
\keyword{mplus}
